+++ src/snowflake/snowpark/modin/plugin/docstrings/series.py
@@ -1249,12 +1249,12 @@ class Series:
 
         Only replace the first NaN element.
 
-        >>> df.fillna(value=values, limit=1)
-            A    B    C    D
-        0  0.0  2.0  2.0  0.0
-        1  3.0  4.0  NaN  1.0
-        2  NaN  1.0  NaN  3.0
-        3  NaN  3.0  NaN  4.0
+        >>> df.fillna(method="ffill", limit=1)
+             A    B   C    D
+        0  NaN  2.0 NaN  0.0
+        1  3.0  4.0 NaN  1.0
+        2  3.0  4.0 NaN  1.0
+        3  NaN  3.0 NaN  4.0
 
         When filling using a DataFrame, replacement happens along
         the same column names and same indices
@@ -1268,6 +1268,10 @@ class Series:
         3  0.0  3.0  0.0  4.0
 
         Note that column D is not affected since it is not present in df2.
+
+        Notes
+        -----
+        `limit` parameter is only supported when using `method` parameter.
         """
 
     @_create_operator_docstring(